Modify Eventlisteners To Output Correct Equations
The Eventlisteners for the buttons with numbers on them were changed
to modify the indexes 1 and 2 of the equation array based upon whether
they were entered before or after an operator.

The Eventlisteners for the buttons with operators on them were changed
to display the operation and then modify the equation array in the case
multiple operations were being done before hitting the equals. This
displays the intermediate operations.

The euals sign event listener was also changed to remove the prior
operator and the second number input, but would save the result of the
prior operation in the case it is used in a following operation. It also
changes the 'afterOperator' variable so that in the case the user begins
typing a new operation it will overwrite the first number.