Modify How Decimals Are Handled
If a decimal is clicked it will either be ignored if there is already a
decimal present, or will be added to the current number and the decimal
will be tracked in the afterDecimal boolean. The equals sign and clear
buttons both change the afterDecimal boolean to false so more decimals
may be added on the new number.